@@ -1,4 +1,4 @@
-class Subscription < ActiveRecord::Base
+class Subscription < ApplicationRecord
 
   self.table_name = 'stripe_subscriptions'
 
